requirements.txt
# === Core Backend Frameworks ===
fastapi==0.116.1
uvicorn==0.33.0

# === AI & Environment ===
openai==0.28.1
python-dotenv==1.0.1
langdetect==1.0.9

# === File Upload Support ===
python-multipart==0.0.9

# === Optional but Useful ===
tqdm==4.67.1
requests==2.32.4

# === JSON and Data Models ===
pydantic==2.10.6
pydantic_core==2.27.2
typing_extensions==4.13.2

# === Chat Logging Support ===
aiofiles==23.2.1

# === Dependency Utilities ===
anyio==4.5.2
sniffio==1.3.1

# === Miscellaneous ===
six==1.17.0
certifi==2025.6.15
charset_normalizer==3.4.2
click==8.1.8
idna==3.10
urllib3==2.5.0

# === Flask Stack (if needed for legacy or mixed use) ===
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.6
greenlet==3.1.1
SQLAlchemy==2.0.41

# === Alembic for DB Migrations ===
alembic==1.14.1

# === Template Rendering Support ===
Jinja2==3.1.6
MarkupSafe==3.0.2
Mako==1.3.10

# === FastAPI Core Dependency ===
starlette==0.44.0

